gRPC/requirements.txt
grpcio
grpcio-tools
protobuf==3.20.0
SQLAlchemy
